+++ packages/atomic/src/components/facets/atomic-date-facet/atomic-date-facet.tsx
@@ -1,4 +1,6 @@
 import {Component, Prop, State, h, Element} from '@stencil/core';
+import dayjs from 'dayjs';
+
 import {
   DateFacet,
   buildDateFacet,
@@ -53,6 +55,7 @@ export class AtomicDateFacet implements InitializableComponent, BaseFacetState {
   public strings: I18nState = {
     clear: () => this.bindings.i18n.t('clear'),
     facetValue: (variables) => this.bindings.i18n.t('facetValue', variables),
+    to: (variables) => this.bindings.i18n.t('to', variables),
   };
 
   @State() public isExpanded = false;
@@ -69,6 +72,10 @@ export class AtomicDateFacet implements InitializableComponent, BaseFacetState {
    * Whether or not the index should automatically generate options for the facet
    */
   @Prop() public generateAutomaticRanges = true;
+  /**
+   * The format that the date will be displayed in. See https://day.js.org/docs/en/display/format for formatting details.
+   */
+  @Prop() public dateFormat = 'DD/MM/YYYY';
 
   public buildOptions() {
     const options = Array.from(this.host.querySelectorAll('atomic-date-range'));
@@ -87,7 +94,10 @@ export class AtomicDateFacet implements InitializableComponent, BaseFacetState {
     this.strings[this.label] = () => this.bindings.i18n.t(this.label);
     this.facet = buildDateFacet(this.bindings.engine, {options});
     this.facetId = this.facet.state.facetId;
-    this.bindings.store.state.facetLabels[this.facetId] = this.label;
+    this.bindings.store.state.facets[this.facetId] = {
+      label: this.label,
+      formatting: this.dateFormat,
+    };
   }
 
   private get values() {
@@ -101,16 +111,21 @@ export class AtomicDateFacet implements InitializableComponent, BaseFacetState {
 
   private buildListItem(item: DateFacetValue) {
     const isSelected = this.facet.isValueSelected(item);
-
+    const start = dayjs(item.start).format(this.dateFormat);
+    const end = dayjs(item.end).format(this.dateFormat);
+    const value = this.strings.to({start, end});
     return (
       <FacetValue
-        label={` ${item.start}-${item.end}`}
+        label={value}
         isSelected={isSelected}
         numberOfResults={item.numberOfResults}
         facetValueSelected={() => {
           this.facet.toggleSelect(item);
         }}
-        ariaLabel={this.strings.facetValue(item)}
+        ariaLabel={this.strings.facetValue({
+          value,
+          numberOfResults: item.numberOfResults,
+        })}
       />
     );
   }